pyproject.toml
[project]
name = "smolagent"
version = "0.1.0"
description = "Virus Risk Assessment Dashboard"
authors = [{ name = "ngmisl", email = "ngmisl@orbiter.website" }]
dependencies = [
    "smolagents>=0.1.0",
    "duckduckgo-search>=7.5.0",
    "python-dotenv>=1.0.0",
    "litellm>=1.61.20",
    "ruff>=0.9.9",
    "pydantic>=2.10.6",
]
requires-python = ">=3.12"
readme = "README.md"
license = { text = "MIT" }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.rye]
managed = true
dev-dependencies = []

[tool.ruff]
line-length = 200
target-version = "py39"

[tool.ruff.lint]
select = [
    "E", # pycodestyle errors
    "W", # pycodestyle warnings
    "F", # pyflakes
    "I", # isort
    "B", # flake8-bugbear
]

[tool.black]
line-length = 200
target-version = ["py39"]
include = '\.pyi?$'

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
